% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcast.R
\name{broadcast}
\alias{broadcast}
\title{broadcasts a shorter vector-like object into a vector of equal length as a longer vector-like object}
\usage{
broadcast(longer, shorter)
}
\arguments{
\item{longer}{longer vector-like object}

\item{shorter}{shorter vector-like object}
}
\description{
broadcasts a shorter vector-like object into a vector of equal length as a longer vector-like object
}
