% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_length.R
\name{check_length}
\alias{check_length}
\title{This function checks that the vector-like objects x, y are of equal length.}
\usage{
check_length(x, y)
}
\arguments{
\item{x}{vector-like object}

\item{y}{vector-like object}
}
\description{
This function checks that the vector-like objects x, y are of equal length.
}
