% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadcast.R
\name{broadcast}
\alias{broadcast}
\title{broadcasts a shorter vector-like object into a vector of equal length as a longer vector-like object
longer longer vector-like object
shorter shorter vector-like object}
\usage{
broadcast(longer, shorter)
}
\description{
broadcasts a shorter vector-like object into a vector of equal length as a longer vector-like object
longer longer vector-like object
shorter shorter vector-like object
}
