% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipr.R
\name{zipr}
\alias{zipr}
\title{zip two vector-like objects into a dataframe}
\usage{
zipr(x = x, y = x, broadcast = FALSE, fill = FALSE, fillvalue = c(NA))
}
\arguments{
\item{x}{vector-like object}

\item{y}{vector-like object}

\item{broadcast}{defaults to FALSE; if TRUE, shorter sequence is repeated until its length is equal to that of the longer sequence}

\item{fill}{defaults to FALSE; bool for whether fillvalue should be implemented}

\item{fillvalue}{value or sequence of value(s) to fill in shorter sequence until it is the same length as the longer sequence. Defaults to NA}
}
\description{
zip two vector-like objects into a dataframe
}
